+++ core/src/main/java/fr/abes/licencesnationales/core/entities/ip/IpV6.java
@@ -55,7 +55,7 @@ public class IpV6 extends IpEntity implements Serializable {
      * @param commentaires Commentaire libre
      * @throws IpException Si l'IP ne peut pas être décodée ou si elle ne respecte pas les contraintes réseaux
      */
-    public IpV6(Long id, String ip, String commentaires) throws IpException {
+    public IpV6(Integer id, String ip, String commentaires) throws IpException {
         super(id, ip, commentaires);
         // On transforme la chaîne de caractère normée en Objet Java
         this.ipRange = Ipv6Range.from(IpUtils.getIP(ip, IpType.IPV6, 1)).to(IpUtils.getIP(ip, IpType.IPV6, 2));
