+++ core/src/main/java/fr/abes/licencesnationales/core/repository/ip/IpV4Repository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IpV4Repository extends JpaRepository<IpV4, Long> {
+public interface IpV4Repository extends JpaRepository<IpV4, Integer> {
 
 }
